+++ src/controllers/Player.js
@@ -1,4 +1,37 @@
+const { playerAttributes } = require("../helpers/constants/Player");
 const { generateError } = require("../helpers/functions/Error");
+
+exports.addPlayerAttribute = (playerId, attribute, game) => {
+    const player = game.players.find(player => player._id.toString() === playerId);
+    const playerAttribute = playerAttributes.find(playerAttribute => playerAttribute.attribute === attribute);
+    player.attributes.push(playerAttribute);
+};
+
+exports.getVotesResults = (votes, allowTie = false) => {
+    const players = [];
+    for (const vote of votes) {
+        const player = players.find(player => player._id === vote.for._id.toString());
+        if (player) {
+            player.vote++;
+        } else {
+            players.push({ _id: vote.for._id.toString(), vote: 1 });
+        }
+    }
+    const maxVotes = Math.max(...players.map(player => player.vote));
+    const nominatedPlayers = players.filter(player => player.vote === maxVotes);
+    if (nominatedPlayers.length > 1 && !allowTie) {
+        throw generateError("BAD_PLAY", "Tie in votes is not allowed for this action.");
+    }
+    return nominatedPlayers;
+};
+
+exports.checkPlayerMultipleVotes = (votes, players) => {
+    for (const player of players) {
+        if (votes.reduce((acc, vote) => vote.from === player._id.toString() ? ++acc : acc, 0) > 1) {
+            throw generateError("BAD_PLAY", `Player with id "${player._id}" isn't allowed to vote more than once.`);
+        }
+    }
+};
 
 exports.checkVoteTarget = (playerId, players, { action }) => {
     const player = players.find(player => player._id.toString() === playerId);
@@ -24,23 +57,19 @@ exports.checkPlayerAbilityToVote = (playerId, players, { action }) => {
 
 exports.checkVoteStructure = vote => {
     if (!vote.from || !vote.for) {
-        throw generateError("BAD_PLAY", "Bad vote structure");
+        throw generateError("BAD_PLAY", "Bad vote structure.");
     } else if (vote.from === vote.for) {
-        throw generateError("BAD_PLAY", "Vote's source and target can't be the same");
+        throw generateError("BAD_PLAY", "Vote's source and target can't be the same.");
     }
 };
 
 exports.checkVotesSourceAndTarget = (votes, { players }, options) => {
-    for (let vote of votes) {
+    for (const vote of votes) {
         this.checkVoteStructure(vote);
         this.checkPlayerAbilityToVote(vote.from, players, options);
-        this.checkVoteTarget(vote.from, players, options);
-        vote = {
-            from: players.find(player => player._id.toString() === vote.from),
-            to: players.find(player => player._id.toString() === vote.to),
-        };
+        this.checkVoteTarget(vote.for, players, options);
     }
-    // this.checkMultiple
+    this.checkPlayerMultipleVotes(votes, players);
 };
 
 exports.checkAndFillVotes = (votes, game, options) => {
@@ -50,6 +79,10 @@ exports.checkAndFillVotes = (votes, game, options) => {
         throw generateError("BAD_PLAY", "`votes` can't be empty");
     }
     this.checkVotesSourceAndTarget(votes, game, options);
+    for (let i = 0; i < votes.length; i++) {
+        votes[i].from = game.players.find(player => player._id.toString() === votes[i].from);
+        votes[i].for = game.players.find(player => player._id.toString() === votes[i].for);
+    }
 };
 
 exports.mayorPlays = async(play, game) => {
@@ -84,14 +117,17 @@ exports.villagersPlay = async(play, game) => {
     console.log("villagers play");
 };
 
-exports.allElectMayor = async(play, game) => {
+exports.allElectMayor = async(play, game, gameHistoryEntry) => {
     const { votes, action } = play;
     this.checkAndFillVotes(votes, game, { action });
+    const nominatedPlayers = this.getVotesResults(votes);
+    this.addPlayerAttribute(nominatedPlayers[0]._id, "mayor", game);
+    gameHistoryEntry.votes = votes;
 };
 
-exports.allPlay = async(play, game) => {
+exports.allPlay = async(play, game, gameHistoryEntry) => {
     const allActions = {
         "elect-mayor": this.allElectMayor,
     };
-    await allActions[play.action](play, game);
+    await allActions[play.action](play, game, gameHistoryEntry);
 };