+++ src/helpers/constants/Player.js
@@ -1,3 +1,10 @@
-exports.playerAttributes = ["mayor", "eaten", "drank-life-potion", "drank-death-potion", "protected", "raven-marked"];
+exports.playerAttributes = [
+    { attribute: "mayor", source: "all" },
+    { attribute: "eaten", source: "wolves" },
+    { attribute: "drank-life-potion", source: "witch" },
+    { attribute: "drank-death-potion", source: "witch" },
+    { attribute: "protected", source: "protector" },
+    { attribute: "raven-marked", source: "raven" },
+];
 
 exports.playerActions = ["eat", "use-potion", "look", "protect", "shoot", "mark", "elect-mayor", "delegate", "settle-votes", "vote"];